% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BIRF.R
\name{BIRF}
\alias{BIRF}
\title{Balanced iterative random forest (Anaissi, Kennedy, Goyal and Catchpoole, 2013)}
\usage{
BIRF(matrix, class)
}
\arguments{
\item{matrix}{Matrix with features in columns and samples in rows}

\item{class}{vector of classification labels for each sample}
}
\description{
This is a feature selection method that iteratively removes features with zero importance using Random Forest
It is mostly used in the selection module.
}
\examples{
\dontrun{
results <- BIRF(data, vector)
}
}
