% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classframe.R
\docType{data}
\name{classframe}
\alias{classframe}
\title{Classification Table that shows the true classification labels for BRCA samples}
\format{
An object of class \code{data.frame} with 500 rows and 1 columns.
}
\usage{
data(classframe)
}
\description{
Classification Table that shows the true classification labels for BRCA samples
}
\examples{
data(classframe)
}
\keyword{datasets}
