% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clinical.R
\docType{data}
\name{clinical}
\alias{clinical}
\title{Subset of Clinical data from the TCGA-BRCA dataset. Ages were assigned to groups
corresponding to the sample's age range}
\format{
An object of class \code{data.frame} with 500 rows and 8 columns.
}
\usage{
data(clinical)
}
\description{
Subset of Clinical data from the TCGA-BRCA dataset. Ages were assigned to groups
corresponding to the sample's age range
}
\examples{
data(clinical)
}
\keyword{datasets}
