% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confounding.R
\docType{data}
\name{confounding}
\alias{confounding}
\title{Subset of confounding data from the TCGA-BRCA dataset. Ages were assigned to groups
corresponding to the sample's age range}
\format{
An object of class \code{data.frame} with 500 rows and 8 columns.
}
\usage{
data(confounding)
}
\description{
Subset of confounding data from the TCGA-BRCA dataset. Ages were assigned to groups
corresponding to the sample's age range
}
\examples{
data(confounding)
}
\keyword{datasets}
