% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance_between.R
\name{distance_between}
\alias{distance_between}
\title{Distance between Two Dataframes}
\usage{
distance_between(df1, df2)
}
\arguments{
\item{df1}{The first dataframe}

\item{df2}{The second dataframe}
}
\description{
This function takes two dataframes and calculates the Euclidean distances between their columns.
}
