% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance_module.R
\name{distance_module}
\alias{distance_module}
\title{Distance Matrices Module}
\usage{
distance_module(
  CBRMSR,
  categorical.similarity = c("Goodall", "Lin"),
  confounding.type = "categorical",
  feature.weights = TRUE
)
}
\arguments{
\item{CBRMSR}{A CBRMSR object}

\item{categorical.similarity}{Either the Goodall3 or Lin1 algorithms from Boriah, Chandola, and Kumar (2008) can be used}

\item{confounding.type}{Presently, categorical is the only type that's supported but other data types are planned}

\item{feature.weights}{A boolean value of whether you want to use feature weights. The selection module must have been run prior to this module for this option to work.}
}
\description{
This module constructs the distance matrices for the confounding and predictor datasets.
Optionally, you can multiply the predictor distance matrix by the inverse
feature weights.
}
\examples{
\dontrun{
CBRMSR <- distance_module(CBRMSR, categorical.similarity = "Goodall", confounding.type = "categorical", feature.weights = TRUE)
}
}
