% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/folding_module.R
\name{folding_module}
\alias{folding_module}
\title{K Folding Module}
\usage{
folding_module(CBRMSR, Folds = 10, LOOC = FALSE)
}
\arguments{
\item{CBRMSR}{A CBRMSR object}

\item{Folds}{Number of folds (for cross-validation)}

\item{LOOC}{A boolean value of whether to use leave one out cross validation. This feature is not supported yet.}
}
\description{
This function creates relatively equal folds
}
\examples{
\dontrun{
CBRMSR <- folding_module(CBRMSR, folds = 10, LOOC = FALSE)
}
}
