% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knn_dist.R
\name{knn_dist}
\alias{knn_dist}
\title{K-Nearest Neighbor with a Distance Matrix}
\usage{
knn_dist(data, distance_matrix, training.labels, k)
}
\arguments{
\item{data}{The matrix containing the data to classify}

\item{distance_matrix}{The distance matrix}

\item{training.labels}{The classification labels to compare to}

\item{K}{The value of K}
}
\description{
This runs the KNN classification algorithm and accepts a distance matrix as an argument
}
