% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microarray.R
\docType{data}
\name{microarray}
\alias{microarray}
\title{A test-set made from a subset of differentially methylated regions in the TCGA-BRCA dataset}
\format{
An object of class \code{data.frame} with 500 rows and 500 columns.
}
\usage{
data(microarray)
}
\description{
A test-set made from a subset of differentially methylated regions in the TCGA-BRCA dataset
}
\examples{
data(microarray)
}
\keyword{datasets}
