% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selection_module.R
\name{selection_module}
\alias{selection_module}
\title{Feature Selection Module}
\usage{
selection_module(CBRMSR, method = c("BIRF", "rknn"))
}
\arguments{
\item{CBRMSR}{A CBRMSR object}

\item{method}{The method for feature selection. Options are BIRF (Balanced Iterative Random Forest) and rknn (random KNN)}
}
\description{
Feature Selection Module
}
\examples{
\dontrun{
# Feature selection with Balanced Iterative Random Forest
CBRMSR <- selection_module(CBRMSR, method = "BIRF")
# Feature selection with random KNN
CBRMSR <- selection_module(CBRMSR, method = "rknn")
}
}
