% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splitting_module.R
\name{splitting_module}
\alias{splitting_module}
\title{Data Splitting Module}
\usage{
splitting_module(CBRMSR, SplitPercent = 0.75)
}
\arguments{
\item{CBRMSR}{A CBRMSR object}

\item{SplitPercent}{Percentage of the training data (requires a decimal less than 1.0)}
}
\description{
This function splits the data into a training set and a testing set after shuffling the rows
}
\examples{
\dontrun{
CBRMSR <- splitting_module(CBRMSR, SplitPercent = 0.75)
}
}
