% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/two_stage_module.R
\name{two_stage_module}
\alias{two_stage_module}
\title{Two Stage Module}
\usage{
two_stage_module(CBRMSR)
}
\arguments{
\item{CBRMSR}{A CBRMSR object}
}
\description{
This module runs a two-stage process. First, it retrieves samples from similar confounding factors before reducing the pool of
retrieved samples through using similarity among the predictor variables. This is done automatically using a confidence metric. Each
training sample is assigned a confidence value which is the average distance to samples of a different class minus the average distance to samples
of the same class. This value is normalized between 0 and 1. This is the last module that should be run.
}
\examples{
\dontrun{
CBRMSR <- two_stage_module(CBRMSR)
}
}
